Fix to memory corruption issue. Fixes to tests.

+++ parser/ast.cpp
@@ -33,7 +33,7 @@ void AST::addChild(AST *t) {
 
 void AST::giveChildren(AST *p)
 {
-    for(int i = 0; i < (int) m_children.size(); i++) {
+    for(size_t i = 0; i < m_children.size(); i++) {
         p->addChild(m_children.at(i));
     }
     m_children.clear();
@@ -47,7 +47,7 @@ void AST::setChildren(vector<AST *> &newChildren)
 
 void AST::deleteChildren()
 {
-    for(int i = 0; i < (int) m_children.size(); i++) {
+    for(size_t i = 0; i < m_children.size(); i++) {
         m_children.at(i)->deleteChildren();
         delete m_children.at(i);
     }
